split python channel interop support
